补充了装备的删除报错处理，完整的解决了装备部分

+++ app/Repositories/EquipmentInfoRepository.php
@@ -29,6 +29,9 @@ class EquipmentInfoRepository
 
     public function destroy($id) {
         $this->model = $this->getById($id);
+        if ($this->model->equipment()->count()) {
+            return true;
+        }
         $this->model->slots()->detach();
         $this->model->delete();
     }
